chore: make accordin a component that accepts children
